Refactor MAME managed path resolution into shared helper
